feat(detect): YOLOv3 디코드 (박스 0건 → 0.23px)
레벨당 출력이 **한 갈래**라 러너가 "레벨에 못 미친다(cls 3, box 0)" 로 멈추고 있었다.
YOLOv3 는 `na×(5+nc)` 한 텐서에 tx,ty,tw,th,obj,cls 를 다 담는다.

전용 디코더를 넣는다. 다른 계열과 셋이 다르다:
 ① 한 갈래 출력을 채널에서 가른다(ch = a*(5+nc)+k).
 ② 앵커가 scales×ratios 가 아니라 **(w,h) 쌍 목록**(base_sizes)이다.
 ③ objectness 로 **먼저 거른다**(conf_thr=0.005). 안 걸면 후보가 수만 개다.
디코드: 중심 = 앵커중심 + (sigmoid(t)−0.5)·stride, **반폭** = 앵커반폭·exp(t).

⚠️ stride 가 **내림차순**이다(32,16,8). 정렬하면 레벨마다 4배씩 어긋난다 — config 순서 유지.

★ 회귀 하나를 만들었다가 잡았다. `AnchorGenerator` 도 `base_sizes` 를 갖는데 그건
**정수 리스트**(32,64,…)라 (w,h) 쌍으로 순회하면 `TypeError` 로 **retinanet export 가
통째로 죽는다**. 안쪽이 순회 가능할 때만 YOLO 쌍으로 본다 — 이름이 같다고 뜻이 같지 않다.

같이: 기준값 전처리를 float32 로 못 박는다. mean/std 가 float64 로 오는 계열이 있어
conv2d 가 `expected scalar type Double but found Float` 로 죽었다(계열 탓이 아니라 dtype).

실측: yolo 텐서 5.87e-04 · 박스 **0.23px PASS**.
회귀: retinanet·atss·yolox·fcos·rtmdet 텐서 전부 통과.

+++ tools/verify/dense_head/verify_postproc.py
@@ -86,7 +86,11 @@ def mmdet_boxes(cfg, ckpt, image, size, thr, to_rgb):
     im = np.asarray(Image.open(image).convert("RGB").resize((size, size), Image.BILINEAR),
                     dtype=np.float32)
     x = im[:, :, ::-1] if to_rgb else im       # 러너 규약에 맞춘다
-    x = (np.ascontiguousarray(x) - mean) / std
+    # ⚠️ **float32 로 못 박는다.** mean/std 가 float64 로 오는 계열이 있어(YOLOv3)
+    #    나눗셈에서 배열이 double 로 승격되고, conv2d 가
+    #    `expected scalar type Double but found Float` 로 죽는다. 계열 탓처럼 보이지만
+    #    전처리 dtype 문제다.
+    x = ((np.ascontiguousarray(x) - mean) / std).astype(np.float32)
     t = torch.from_numpy(x).permute(2, 0, 1).unsqueeze(0)
 
     meta = {"img_shape": (size, size), "ori_shape": (size, size),
